refactor: move mock components into mock package
