+++ act/actuation/__init__.py
@@ -1,7 +1,7 @@
 # actuation パッケージ
 # コマンドを物理信号（PWM等）に変換・出力する駆動モジュールの実装
 
-from .mock import MockActuation
+from ..mock.actuation import MockActuation
 
 # pwm.pyのインポートは条件付き（実機環境でのみ利用可能）
 try:
